requirements.txt
pygame==2.5.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pandas==2.1.0
Pillow==10.0.0
numpy==1.24.0